pyproject.toml
[tool.poetry]
name = "fade"
version = "0.0.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.10"
s3path = "*"
torch = ">=1.9.0"
torchvision = "*"
regex = "*"
ftfy = "*"
tqdm = "*"
huggingface_hub = "*"
sentencepiece = "*"
protobuf = "*"
timm = "*"
einops = "*"
open_clip_torch = "*"
opencv-python = "*"
matplotlib = "*"
numpy = "*"
requests = "*"
notebook = "^7.1.1"
xgboost = "^2.0.3"
gem-torch = "^1.0.1"
ruff = "^0.6.2"
scikit-learn = "^1.5.1"
black = "^24.8.0"
click = "^8.1.7"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
